SpinLock: rework using Go's fair algorithm
The previous implementation was spinning a 100 times on an atomic swap
before yielding to another thread, which was slower than the private
`Crystal::SpinLock`.

This new version implement's Go's futex algorithm, where we spin 4 times
trying to swap the atomic while easing the cpu between each attempts (to
avoid burning) before yielding to another thread.

The futex sleep/wakeup wasn't ported as the CAS operations (to keep the
SLEEPING state) and the actual futex were slower in both a naive
benchmark, where all threads are trying to lock, and another benchmark
where only half the threads are trying to lock while the other half are
busy doing the same CPU bound operation but without the lock.

The new algorithm proved faster on high contention, and perfectly stable
once we have more threads than available CPU cores, which the exception
of MT:8 for which `Crystal::SpinLock` performed better on the all
threads fight for the lock benchmark (but not the distributed
lock/nolock one).